changed price from double to int

+++ src/main/java/com/vahundos/model/MenuMeal.java
@@ -12,7 +12,7 @@ public class MenuMeal {
     private Meal meal;
 
     @Column(name = "price")
-    private double price;
+    private int price;
 
     public Meal getMeal() {
         return meal;
@@ -26,7 +26,7 @@ public class MenuMeal {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
